Revise component name to ProfileSettings and create radio buttons for pronouns
